quick refactor of directory structure, was bugging me
