+++ src/main/java/com/controllerface/bvge/gl/renderers/BoundingBoxRenderer.java
@@ -1,6 +1,5 @@
-package com.controllerface.bvge.ecs.systems.renderers;
+package com.controllerface.bvge.gl.renderers;
 
-import com.controllerface.bvge.Main;
 import com.controllerface.bvge.cl.GPU;
 import com.controllerface.bvge.ecs.ECS;
 import com.controllerface.bvge.ecs.systems.GameSystem;
@@ -9,46 +8,58 @@ import com.controllerface.bvge.util.Assets;
 import com.controllerface.bvge.util.Constants;
 import com.controllerface.bvge.window.Window;
 
+import java.util.HashMap;
+import java.util.Map;
+import java.util.function.Function;
+
 import static org.lwjgl.opengl.GL11C.GL_FLOAT;
+import static org.lwjgl.opengl.GL11C.GL_LINE_LOOP;
 import static org.lwjgl.opengl.GL15C.GL_ARRAY_BUFFER;
 import static org.lwjgl.opengl.GL15C.GL_DYNAMIC_DRAW;
-import static org.lwjgl.opengl.GL15C.GL_LINES;
 import static org.lwjgl.opengl.GL15C.glBindBuffer;
 import static org.lwjgl.opengl.GL15C.glBufferData;
-import static org.lwjgl.opengl.GL15C.glDrawArrays;
 import static org.lwjgl.opengl.GL15C.glGenBuffers;
+import static org.lwjgl.opengl.GL15C.glMultiDrawArrays;
 import static org.lwjgl.opengl.GL20C.glDisableVertexAttribArray;
 import static org.lwjgl.opengl.GL20C.glEnableVertexAttribArray;
 import static org.lwjgl.opengl.GL20C.glVertexAttribPointer;
 import static org.lwjgl.opengl.GL30C.glBindVertexArray;
 import static org.lwjgl.opengl.GL30C.glGenVertexArrays;
 
-/**
- * Renders physics edge constraints. All defined edges are rendered as lines.
- */
-public class EdgeRenderer extends GameSystem
+public class BoundingBoxRenderer extends GameSystem
 {
     private static final int VERTEX_SIZE = 2;
-    private static final int VERTS_PER_EDGE = 2;
+    private static final int VERTS_PER_BOX = 4;
     private static final int VERTEX_SIZE_BYTES = VERTEX_SIZE * Float.BYTES;
-    private static final int BATCH_VERTEX_COUNT = Constants.Rendering.MAX_BATCH_SIZE * VERTS_PER_EDGE * VERTEX_SIZE;
+    private static final int BATCH_VERTEX_COUNT = Constants.Rendering.MAX_BATCH_SIZE * VERTS_PER_BOX * VERTEX_SIZE;
     private static final int BATCH_BUFFER_SIZE = BATCH_VERTEX_COUNT * Float.BYTES;
-
-    private static final int FLAG_SIZE = 1;
-    private static final int FLAGS_PER_EDGE = 2;
-    private static final int FLAG_SIZE_BYTES = FLAG_SIZE * Float.BYTES;
-    private static final int BATCH_FLAG_COUNT = Constants.Rendering.MAX_BATCH_SIZE * FLAGS_PER_EDGE * FLAG_SIZE;
-    private static final int BATCH_FLAG_SIZE = BATCH_FLAG_COUNT * Float.BYTES;
-
     private final AbstractShader shader;
     private int vao_id;
-    private int edge_vbo;
-    private int flag_vbo;
+    private int bounding_box_vbo;
+    private final int[] offsets = new int[Constants.Rendering.MAX_BATCH_SIZE];
+    private final int[] counts = new int[Constants.Rendering.MAX_BATCH_SIZE];
+    private final Map<Integer, BatchCounts> batch_cache = new HashMap<>();
 
-    public EdgeRenderer(ECS ecs)
+    private record BatchCounts(int[] offsets, int[] counts) { }
+
+    private final Function<Integer, BatchCounts> count_batch = (count) ->
+    {
+        int[] counts = new int[count];
+        int[] offsets = new int[count];
+        System.arraycopy(this.counts, 0, counts, 0, count);
+        System.arraycopy(this.offsets, 0, offsets, 0, count);
+        return new BatchCounts(offsets, counts);
+    };
+
+    public BoundingBoxRenderer(ECS ecs)
     {
         super(ecs);
-        this.shader = Assets.load_shader("object_outline.glsl");
+        this.shader = Assets.load_shader("bounding_outline.glsl");
+        for (int i = 0; i < Constants.Rendering.MAX_BATCH_SIZE; i++)
+        {
+            offsets[i] = i * 4;
+            counts[i] = 4;
+        }
         init();
     }
 
@@ -57,17 +68,11 @@ public class EdgeRenderer extends GameSystem
         vao_id = glGenVertexArrays();
         glBindVertexArray(vao_id);
 
-        edge_vbo = glGenBuffers();
-        glBindBuffer(GL_ARRAY_BUFFER, edge_vbo);
+        bounding_box_vbo = glGenBuffers();
+        glBindBuffer(GL_ARRAY_BUFFER, bounding_box_vbo);
         glBufferData(GL_ARRAY_BUFFER, BATCH_BUFFER_SIZE, GL_DYNAMIC_DRAW);
         glVertexAttribPointer(0, VERTEX_SIZE, GL_FLOAT, false, VERTEX_SIZE_BYTES, 0);
-        GPU.share_memory(edge_vbo);
-
-        flag_vbo = glGenBuffers();
-        glBindBuffer(GL_ARRAY_BUFFER, flag_vbo);
-        glBufferData(GL_ARRAY_BUFFER, BATCH_FLAG_SIZE, GL_DYNAMIC_DRAW);
-        glVertexAttribPointer(1, FLAG_SIZE, GL_FLOAT, false, FLAG_SIZE_BYTES, 0);
-        GPU.share_memory(flag_vbo);
+        GPU.share_memory(bounding_box_vbo);
 
         glBindBuffer(GL_ARRAY_BUFFER, 0);
         glBindVertexArray(0);
@@ -82,19 +87,18 @@ public class EdgeRenderer extends GameSystem
         shader.uploadMat4f("uVP", Window.get().camera().get_uVP());
 
         glEnableVertexAttribArray(0);
-        glEnableVertexAttribArray(1);
 
         int offset = 0;
-        for (int remaining = GPU.Memory.next_edge(); remaining > 0; remaining -= Constants.Rendering.MAX_BATCH_SIZE)
+        for (int remaining = GPU.Memory.next_hull(); remaining > 0; remaining -= Constants.Rendering.MAX_BATCH_SIZE)
         {
             int count = Math.min(Constants.Rendering.MAX_BATCH_SIZE, remaining);
-            GPU.GL_edges(edge_vbo, flag_vbo, offset, count);
-            glDrawArrays(GL_LINES, 0, count * 2);
+            var batch_counts = batch_cache.computeIfAbsent(count, count_batch);
+            GPU.GL_bounds(bounding_box_vbo, offset, count);
+            glMultiDrawArrays(GL_LINE_LOOP, batch_counts.offsets, batch_counts.counts);
             offset += count;
         }
 
         glDisableVertexAttribArray(0);
-        glDisableVertexAttribArray(1);
         glBindVertexArray(0);
 
         shader.detach();
