+++ src/main/java/com/controllerface/bvge/gl/renderers/CrateRenderer.java
@@ -1,4 +1,4 @@
-package com.controllerface.bvge.ecs.systems.renderers;
+package com.controllerface.bvge.gl.renderers;
 
 import com.controllerface.bvge.cl.GPU;
 import com.controllerface.bvge.cl.HullIndexData;
