+++ src/main/java/com/controllerface/bvge/window/Window.java
@@ -74,8 +74,6 @@ public class Window
 
     public void run()
     {
-        System.out.println("LWJGL version: " + Version.getVersion());
-
         glfwShowWindow(glfwWindow);
         loop();
 
@@ -204,6 +202,7 @@ public class Window
         // set current cursor
         glfwSetCursor(glfwWindow, cursorID);
 
+        System.out.println("LWJGL version: " + Version.getVersion());
 
         System.out.println("\n-------- OPEN GL DEVICE -----------");
         System.out.println(glGetString(GL_VENDOR));
