+++ src/main/resources/gl/shaders/bone_shader.glsl
