+++ src/main/resources/gl/shaders/bounding_outline.glsl
