+++ src/main/resources/gl/shaders/box_model.glsl
