+++ src/main/resources/gl/shaders/circle_shader.glsl
