+++ src/main/resources/gl/shaders/object_outline.glsl
