+++ src/main/resources/gl/shaders/point_shader.glsl
