+++ src/main/resources/gl/shaders/poly_model.glsl
